Move macro_rules diagnostics to diagnostics module
